Add files via upload
You can set the date for today by the spreadsheet function, Today(), but you can see  the date is not fixed when you open the file in another day. It is customized to check if there is already input date, and it will put the date of today if there is no input date in the cells.